+++ src/sparseml/pytorch/sparsification/quantization/__init__.py
@@ -15,3 +15,4 @@
 # limitations under the License.
 
 from .modifier_quantization import *
+from .utils import *
