+++ src/sparseml/pytorch/sparsification/quantization/modifier_quantization.py
@@ -47,7 +47,7 @@ except Exception:
 from sparseml.optim import BaseModifier, ModifierProp
 from sparseml.pytorch.optim.modifier import PyTorchModifierYAML, ScheduledModifier
 from sparseml.pytorch.utils import BaseLogger, tensors_module_forward, tensors_to_device
-from sparseml.pytorch.utils.quantization import (
+from sparseml.pytorch.sparsification.quantization import (
     add_quant_dequant,
     configure_module_default_qconfigs,
     configure_module_qat_wrappers,
